+++ MergableMigrations.Specification/Implementation/MigrationMemento.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Collections.Immutable;
+using System.Numerics;
 
 namespace MergableMigrations.Specification.Implementation
 {
@@ -7,10 +8,10 @@ namespace MergableMigrations.Specification.Implementation
     {
         public string Type { get; }
         public ImmutableDictionary<string, string> Attributes { get; }
-        public int HashCode { get; }
-        public ImmutableList<int> Prerequisites { get; }
+        public BigInteger HashCode { get; }
+        public ImmutableList<BigInteger> Prerequisites { get; }
 
-        public MigrationMemento(string type, IDictionary<string, string> attributes, int hashCode, IEnumerable<int> prerequisites)
+        public MigrationMemento(string type, IDictionary<string, string> attributes, BigInteger hashCode, IEnumerable<BigInteger> prerequisites)
         {
             Type = type;
             Attributes = attributes.ToImmutableDictionary();
